+++ R/fs_perms.R
@@ -87,12 +87,12 @@ format.fs_perms <- function(x, ...) {
 as.character.fs_perms <- format.fs_perms
 
 #' @export
-`[.fs_perms` <- function(x, i) {
+`[.fs_perms` <- function(x, i, ...) {
   new_fs_perms(NextMethod("["))
 }
 
 #' @export
-`[[.fs_perms` <- function(x, i) {
+`[[.fs_perms` <- function(x, i, ...) {
   new_fs_perms(NextMethod("[["))
 }