+++ R/utils.R
@@ -125,3 +125,15 @@ fs_error <- function(msg, class = "invalid_argument") {
 lengths <- function(x) {
   vapply(x, length, integer(1))
 }
+
+as_tibble <- function(x) {
+  if (is_installed("tibble")) {
+    tibble::as_tibble(x)
+  } else {
+    x
+  }
+}
+
+is_installed <- function(pkg) {
+  isTRUE(requireNamespace(pkg, quietly = TRUE))
+}